Stop connect to DB on every request in SQLite DApp
Instead of opening a new connection on every request, SQLite DApp now
connects on startup, and simply retrieves a cursor to execute the
specified SQL statement. This is nice to illustrate that memory state
is preserved between inputs, and also that heavier initialization
procedures can be done beforehand instead of on every input.